# swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DepressoSpeech – Unified API Documentation",
    "description": "Comprehensive API documentation for the DepressoSpeech Depression Severity Prediction System.\n\n## Services\n\n| Service | Port | Description |\n|---------|------|-------------|\n| **MindScope Backend** | 8000 | Core application API – authentication, assessments, file management, doctor/admin dashboards |\n| **DepressoSpeech ML Model** | 8001 | ML inference API – PHQ-8 prediction from audio using eGeMAPS + MFCC + SBERT features |\n| **Frontend (Vite)** | 5173 | React SPA – patient/doctor/admin user interface |\n| **Swagger Docs** | 8080 | This documentation page |\n\n## Authentication\n- **Backend API**: JWT Bearer tokens via `/api/v1/auth/login`\n- **ML Model API**: Optional API key via `X-API-Key` header\n- **Admin**: Separate admin login via `/api/v1/auth/admin/login`\n\n## Architecture\nThe system follows a microservices pattern:\n1. Frontend sends requests to Backend (proxied via Vite in dev)\n2. Backend orchestrates business logic, stores data in SQLite\n3. Backend delegates ML inference to the Model service\n4. Model service processes audio and returns PHQ-8 predictions",
    "version": "2.0.0",
    "contact": {
      "name": "DepressoSpeech Team – KMIT",
      "email": "koushik@kmit.in"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
...
    },
    {
...
    }
  ],
...
  "paths": {
...
    "/doctor/dashboard/summary": {
      "get": {
        "tags": ["Doctor Dashboard"],
        "summary": "Doctor dashboard summary",
        "description": "Aggregate statistics: total patients, assessments, high/low risk cases, severity breakdown.",
        "operationId": "doctorSummary",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Dashboard summary data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DoctorSummaryResponse" }
              }
            }
          },
          "403": { "description": "Not a doctor" }
        }
      }
    },
    "/doctor/dashboard/alerts": {
      "get": {
        "tags": ["Doctor Dashboard"],
        "summary": "High-risk alerts",
        "description": "Recent assessments flagged as Severe or Moderately Severe.",
        "operationId": "doctorAlerts",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "severity", "in": "query", "schema": { "type": "array", "items": { "type": "string" } }, "description": "Filter by severity levels" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 12, "minimum": 1, "maximum": 50 } }
        ],
        "responses": {
          "200": { "description": "List of high-risk alerts" }
        }
      }
    },
    "/doctor/dashboard/patient-trends": {
      "get": {
        "tags": ["Doctor Dashboard"],
        "summary": "Patient trends",
        "description": "Assessment score trends per patient over time.",
        "operationId": "patientTrends",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "patientId", "in": "query", "schema": { "type": "string" }, "description": "Optional specific patient ID" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 } }
        ],
        "responses": {
          "200": { "description": "Patient trend data" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from `/api/v1/auth/login`"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Optional API key for ML model (set via DEPRESSO_API_KEY env var)"
      }
    },
    "schemas": {
...
      "DoctorSummaryResponse": {
        "type": "object",
        "properties": {
          "totals": {
            "type": "object",
            "properties": {
              "patients": { "type": "integer" },
              "assessments": { "type": "integer" },
              "highRiskCases": { "type": "integer" },
              "lowRiskCases": { "type": "integer" }
            }
          },
          "severityBreakdown": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "severity": { "type": "string" },
                "count": { "type": "integer" }
              }
            }
          }
        }
      },
...
    }
  }
}